Move stopLayerStopType to core-utils types
