+++ packages/core-utils/src/types.js
@@ -350,6 +350,13 @@ export const transitIndexStopWithRoutes = PropTypes.shape({
   )
 });
 
+export const stopLayerStopType = PropTypes.shape({
+  id: PropTypes.string.isRequired,
+  name: PropTypes.string.isRequired,
+  lat: PropTypes.number.isRequired,
+  lon: PropTypes.number.isRequired
+});
+
 const transitivePlaceType = PropTypes.shape({
   place_id: PropTypes.string.isRequired,
   type: PropTypes.string.isRequired
